We don't need to do all the other checks.
We first check if player moved enough for us to care. We can't do >>4 like Factions since plot size is variable, and not always cut-off on chunk, so we are checking integer values instead. This allows for basic fidgeting and flying vertically up or down without triggering it. If that is the case, we drop straight to checking whether or not they've changed plot -- relatively light weight string comparison.

This should help make the onPlayerMove hook lighter than necessary.